backend/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
langchain==0.1.5
openai==1.12.0
faiss-cpu==1.7.4
chromadb==0.4.22
python-multipart==0.0.9
tiktoken==0.6.0 